refactor(appkit): extract pure agents-plugin helpers into sibling modules
Step 1 of splitting the ~2.5k-line agents plugin. Moves module-scope pure
functions/constants out of agents.ts verbatim (behavior-preserving):
- approval.ts            requiresApproval
- prompt.ts              composePromptForAgent
- builtin-tools.ts       LOAD_SKILL_TOOL_DEF, READ_SKILL_FILE_TOOL_DEF
- adapter-extensions.ts  buildAdapterExtensions, supervisorToolDescription, warnOnCapabilityMismatch

agents.ts: 2512 -> 2331 lines. typecheck + 394 agent tests green.
